update some code based on sonar issues
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1896504 13f79535-47bb-0310-9956-ffa450edef68

+++ poi/src/main/java/org/apache/poi/ss/formula/atp/XMatchFunction.java
@@ -94,7 +94,7 @@ final class XMatchFunction implements FreeRefFunction {
                 vector = LookupUtils.createRowVector(tableArray, 0);
             }
             int matchedIdx = LookupUtils.xlookupIndexOfValue(lookupValue, vector, matchMode, searchMode);
-            return new NumberEval(matchedIdx + 1);
+            return new NumberEval((double)matchedIdx + 1);
         } catch (EvaluationException e) {
             return e.getErrorEval();
         }
